% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oosbase.R
\name{oos_dt_get_file}
\alias{oos_dt_get_file}
\title{get a file with data table library . Fread function}
\usage{
oos_dt_get_file(credentials, container, file_name, cols)
}
\arguments{
\item{credentials}{List Object returned from a call to oos_autheticate}

\item{container}{Full container name where the file resides.}

\item{file_name}{the name of the file to be removed}
}
\value{
returns any errors.  NULL means success.
}
\description{
get a file with data table library . Fread function
}
\examples{
#oos_dt_get_file(my_credentials,"sales","sales.csv", cols )
}
